Ignore Numba warnings and pin supported versions
Ignores recent Numba warnings that are related to how umap-learn uses
Numba and outside of our control as users of umap-learn. Following
recommendations from the Numba deprecation guide [1], we explicitly
ignore the `NumbaDeprecationWarning` and pin the maximum Numba version
that will support the current umap-learn usage. We also pin umap-learn
to a maximum minor version in preparation for a potential new minor
version that addresses the Numba issue [2]

[1] https://numba.readthedocs.io/en/stable/reference/deprecation.html#deprecation-of-object-mode-fall-back-behaviour-when-using-jit
[2] https://github.com/lmcinnes/umap/issues/1004