Fix ghosting of existing variable by new variable
Fixes a bug in the embed logic where renaming a previous `cluster_data`
variable to `embedding_parameters` accidentally interfered with an
existing variable of the same name and similar purpose. The fix renames
the new variable to `external_embedding_parameters` to disambiguate its
purpose and source from the original `embedding_parameters` variable.