Use number of components defined previously
Changes logic for number of components to always use the `n_components`
variable that either uses the command line argument provided by the user
or an external embedding parameters file. If we only used the former, we
would fail to properly setup number of components when users provide an
external embedding parameters file.