Use cluster label attribute as full column name
Instead of treating the new cluster label attribute argument value as a
suffix for a string like "label_<argument>", uses the argument value as
the full column name for the label. This change provides users with more
flexibility in how their output data are formatted, and it still allows
them to follow the "label_<argument>" pattern.