Fix UMAP multiprocessing user warning
Fixes a `UserWarning` produced by UMAP when providing a random seed
value. By default, UMAP tries to use all available CPUs (`n_jobs=-1`)
but multiprocessing only works when the random seed is not set. This
commit sets the number of CPUs to 1, since we need the reproducibility
provided by setting the random seed and we don't need the warning.