Convert boxplots to base matplotlib
Removes original "scatterplot" function in favor of working with
compressed distance matrix formats for both genetic and Euclidean
distances, groups Euclidean distances into arrays by genetic distance to
prepare for plotting with base matplotlib, and then plots boxplots with
base matplotlib. Importantly, this implementation creates a boxplot
array for each integer value in the range of observed genetic distances
including those intervening distances that were not observed. This
implementation change allows there to be gaps between genetic distances
on the x-axis in a way that our previous plots did not allow and which I
think resulted in some genetic distance boxes getting placed in the
wrong position on the x axis. This implementation also explicitly
positions and labels each box on the x axis to match the observed
range of genetic distance values.

This commit makes a couple of other minor aesthetic improvements to the
original plots including plotting of axis ticks every 5 distance units,
using the matplotlib "tight layout" to remove outer whitespace in the
figure, and plotting of the linear fit between distance axes behind the
boxplots as a reference for the equation shown in the title.