Allow Python stack traces on error
Instead of catching exceptions for each subcommand and printing the
exception error alone, just let the exception happen as normal and print
the full stack trace to the screen. The error message by itself is not
useful for debugging issues, so I always end up having to temporarily
disable the exception catching.