Sort distance inputs to embed
Sorts each distance matrix input alphabetically by sequence name (rows
and columns) on load to ensure that the sum of all matrices correctly
represents the same pairs of sequences. Adds an assertion to confirm
that sequence names are identical for each distance matrix input. While
we're improving this function, we also create an index of the sequence
names for the final distance matrix data frame, so we don't have to set
it later from alignment names. Adds a functional test to confirm that
this behavior works as expected.