Ensure distance matrices are sorted and match
Adds/updates tests to ensure that distance matrices provided by the user
or created on the fly from alignments are sorted and have the same
samples. Refactors the logic for calculating distances from an alignment
into its own function, so we don't have to copy the same code for
pathogen-distance and pathogen-embed.