Fix OMP warnings in UMAP tests
Explicitly sets the threading layer for Numba to a simple implementation
that should always be available. UMAP uses Numba internally and, on some
systems, the Numba threading layer can get set to an OpenMP (OMP)
backend. Since Numba internally calls OMP with a function that is
deprecated, OMP emits a deprecation warning. This warning is especially
useless since we force UMAP to run with a single CPU. The main effect of
this warning was to break our functional tests.

For more details, see this Numba issue comment on GitHub:
https://github.com/numba/numba/issues/5275#issuecomment-1002935625