Fix a bug plotting pairwise distances for PCA
Fixes a bug where requesting a pairwise distance plot with the PCA
embedding crashed the command because the genetic distance matrix was
missing. This commit updates the logic to determine whether the distance
matrix needs to be created on the fly to include the case when the user
has requested the pairwise distance figure.