Automatically pick SVD algorithm for PCA from data
Let scikit-learn automatically pick the SVD algorithm to use for PCA
based on the given data instead of hardcoding the "full" solver. Fixes a
bug that can occur with random virus datasets where the full SVD solver
cannot converge but the randomized solver can. The scikit-learn PCA
class picks from the full, arpack, and randomized solvers based on the
shape of the input data matrix [1] which makes PCA more robust to these
kinds of convergence issues.

[1] https://scikit-learn.org/stable/modules/generated/sklearn.decomposition.PCA.html#pca